+++ src/model/ops.py
@@ -1,3 +1,4 @@
+import math
 import re
 import numpy as np
 import torch
@@ -6,8 +7,8 @@ import torch.nn as nn
 import torchaudio
 import sys
 from torch.autograd import Variable
-import math
 
+from .modules import initialize_filterbank
 
 class Conv_2d(nn.Module):
     def __init__(self, input_channels, output_channels, shape=3, stride=1, pooling=2):
@@ -20,7 +21,7 @@ class Conv_2d(nn.Module):
         out = self.mp(self.relu(self.bn(self.conv(x))))
         return out
 
-class Res_2d_mp(nn.Module):
+class Res2Dmp(nn.Module):
     def __init__(self, input_channels, output_channels, pooling=2):
         super(Res_2d_mp, self).__init__()
         self.conv_1 = nn.Conv2d(input_channels, output_channels, 3, padding=1)
@@ -36,18 +37,33 @@ class Res_2d_mp(nn.Module):
         return out
 
 class HarmonicSTFT(nn.Module):
+    """HarmonicSTFT class"""
     def __init__(self,
-                 sample_rate=16000,
-                 n_fft=513,
-                 win_length=None,
-                 hop_length=None,
-                 pad=0,
-                 power=2,
-                 normalized=False,
-                 n_harmonic=6,
-                 semitone_scale=2,
-                 bw_Q=1.0,
-                 learn_bw=None):
+                 sample_rate:float,
+                 n_fft:float,
+                 win_length:int,
+                 hop_length:int,
+                 pad:int,
+                 power:int,
+                 normalized:bool,
+                 n_harmonic:int,
+                 semitone_scale:int,
+                 bw_Q:int,
+                 learn_bw:str)
+        """Instantiating HarmonicCNN class
+            Args:
+                sample_rate(float) : sampling rate
+                n_fft(float) : number of fft bin
+                win_length:int,
+                hop_length:int,
+                pad(int) : pad value of STFT,
+                power(int) : Power of STFT,
+                normalized(bool) : Normalize STFT,
+                n_harmonic(int) : the number of n_harmonic
+                semitone_scale(int) : a half step or a half tone, is the smallest musical interval
+                bw_Q(int) : bandwidths initalizae,
+                learn_bw(str) : learnable bandwidths
+        """
         super(HarmonicSTFT, self).__init__()
 
         # Parameters
@@ -76,7 +92,10 @@ class HarmonicSTFT(nn.Module):
             self.bw_Q = torch.tensor(np.array([bw_Q]).astype('float32'))
 
     def get_harmonic_fb(self):
-        # bandwidth
+        """get harmonic filter bank
+        Returns:
+            based on upslope and downslop, get triangle audiotory filter
+        """
         bw = (self.bw_alpha * self.f0 + self.bw_beta) / self.bw_Q
         bw = bw.unsqueeze(0) # (1, n_band)
         f0 = self.f0.unsqueeze(0) # (1, n_band)
@@ -87,7 +106,7 @@ class HarmonicSTFT(nn.Module):
         fb = torch.max(self.zero, torch.min(down_slope, up_slope))
         return fb
 
-    def to_device(self, device, n_bins):
+    def to_device(self, device: torch.cuda, n_bins: torch.Tensor):
         self.f0 = self.f0.to(device)
         self.bw_Q = self.bw_Q.to(device)
         # fft bins
@@ -96,186 +115,16 @@ class HarmonicSTFT(nn.Module):
         self.zero = torch.zeros(1)
         self.zero = self.zero.to(device)
 
-    def forward(self, waveform):
+    def forward(self, waveform: torch.Tensor) -> torch.Tensor:
         # stft
         spectrogram = self.spec(waveform)
-
         # to device
         self.to_device(waveform.device, spectrogram.size(1))
-
         # triangle filter
         harmonic_fb = self.get_harmonic_fb()
         harmonic_spec = torch.matmul(spectrogram.transpose(1, 2), harmonic_fb).transpose(1, 2)
-
         # (batch, channel, length) -> (batch, harmonic, f0, length)
         b, c, l = harmonic_spec.size()
         harmonic_spec = harmonic_spec.view(b, self.n_harmonic, self.level, l)
-
-        # amplitude to db
         harmonic_spec = self.amplitude_to_db(harmonic_spec)
         return harmonic_spec
-
-
-# Modules for harmonic filters
-
-
-def initialize_filterbank(sample_rate, n_harmonic, semitone_scale):
-    # MIDI
-    # lowest note
-    low_midi = note_to_midi('C1')
-
-    # highest note
-    high_note = hz_to_note(sample_rate / (2 * n_harmonic))
-    high_midi = note_to_midi(high_note)
-
-    # number of scales
-    level = (high_midi - low_midi) * semitone_scale
-    midi = np.linspace(low_midi, high_midi, level + 1)
-    hz = midi_to_hz(midi[:-1])
-
-    # stack harmonics
-    harmonic_hz = []
-    for i in range(n_harmonic):
-        harmonic_hz = np.concatenate((harmonic_hz, hz * (i+1)))
-
-    return harmonic_hz, level
-    
-def hz_to_midi(hz):
-    return 12 * (torch.log2(hz) - np.log2(440.0)) + 69
-
-def midi_to_hz(midi):
-    return 440.0 * (2.0 ** ((midi - 69.0)/12.0))
-
-def note_to_midi(note):
-    pitch_map = {"C": 0, "D": 2, "E": 4, "F": 5, "G": 7, "A": 9, "B": 11}
-    acc_map = {"#": 1,"": 0,"b": -1,"!": -1,"♯": 1,"𝄪": 2,"♭": -1,"𝄫": -2,"♮": 0}
-
-    match = re.match(
-        r"^(?P<note>[A-Ga-g])"
-        r"(?P<accidental>[#♯𝄪b!♭𝄫♮]*)"
-        r"(?P<octave>[+-]?\d+)?"
-        r"(?P<cents>[+-]\d+)?$",
-        note)
-    pitch = match.group("note").upper()
-    offset = np.sum([acc_map[o] for o in match.group("accidental")])
-    octave = match.group("octave")
-    cents = match.group("cents")
-
-    if not octave:
-        octave = 0
-    else:
-        octave = int(octave)
-    if not cents:
-        cents = 0
-    else:
-        cents = int(cents) * 1e-2
-
-    note_value = 12 * (octave + 1) + pitch_map[pitch] + offset + cents
-    return note_value
-
-def midi_to_note(midi, octave=True, cents=False, key="C:maj", unicode=True):
-
-    note_map = key_to_notes(key=key, unicode=unicode)
-
-    note_num = int(np.round(midi))
-    note_cents = int(100 * np.around(midi - note_num, 2))
-
-    note = note_map[note_num % 12]
-
-    if octave:
-        note = "{:s}{:0d}".format(note, int(note_num / 12) - 1)
-    if cents:
-        note = "{:s}{:+02d}".format(note, note_cents)
-
-    return note
-
-def hz_to_note(hz):
-    return midi_to_note(hz_to_midi(hz))
-
-def key_to_notes(key, unicode=True):
-    # Parse the key signature
-    match = re.match(
-        r"^(?P<tonic>[A-Ga-g])"
-        r"(?P<accidental>[#♯b!♭]?)"
-        r":(?P<scale>(maj|min)(or)?)$",
-        key,
-    )
-    if not match:
-        raise ParameterError("Improper key format: {:s}".format(key))
-    pitch_map = {"C": 0, "D": 2, "E": 4, "F": 5, "G": 7, "A": 9, "B": 11}
-    acc_map = {"#": 1, "": 0, "b": -1, "!": -1, "♯": 1, "♭": -1}
-
-    tonic = match.group("tonic").upper()
-    accidental = match.group("accidental")
-    offset = acc_map[accidental]
-
-    scale = match.group("scale")[:3].lower()
-    major = scale == "maj"
-
-    if major:
-        tonic_number = ((pitch_map[tonic] + offset) * 7) % 12
-    else:
-        tonic_number = ((pitch_map[tonic] + offset) * 7 + 9) % 12
-
-    if offset < 0:
-        use_sharps = False
-
-    elif offset > 0:
-        use_sharps = True
-
-    elif 0 <= tonic_number < 6:
-        use_sharps = True
-
-    elif tonic_number > 6:
-        use_sharps = False
-
-    # Basic note sequences for simple keys
-    notes_sharp = ["C", "C♯", "D", "D♯", "E", "F", "F♯", "G", "G♯", "A", "A♯", "B"]
-    notes_flat = ["C", "D♭", "D", "E♭", "E", "F", "G♭", "G", "A♭", "A", "B♭", "B"]
-
-    sharp_corrections = [
-        (5, "E♯"),
-        (0, "B♯"),
-        (7, "F𝄪"),
-        (2, "C𝄪"),
-        (9, "G𝄪"),
-        (4, "D𝄪"),
-        (11, "A𝄪"),
-    ]
-
-    flat_corrections = [
-        (11, "C♭"),
-        (4, "F♭"),
-        (9, "B𝄫"),
-        (2, "E𝄫"),
-        (7, "A𝄫"),
-        (0, "D𝄫"),
-    ]  # last would be (5, 'G𝄫')
-
-    # Apply a mod-12 correction to distinguish B#:maj from C:maj
-    n_sharps = tonic_number
-    if tonic_number == 0 and tonic == "B":
-        n_sharps = 12
-
-    if use_sharps:
-        # This will only execute if n_sharps >= 6
-        for n in range(0, n_sharps - 6 + 1):
-            index, name = sharp_corrections[n]
-            notes_sharp[index] = name
-
-        notes = notes_sharp
-    else:
-        n_flats = (12 - tonic_number) % 12
-
-        # This will only execute if tonic_number <= 6
-        for n in range(0, n_flats - 6 + 1):
-            index, name = flat_corrections[n]
-            notes_flat[index] = name
-
-        notes = notes_flat
-
-    # Finally, apply any unicode down-translation if necessary
-    if not unicode:
-        translations = str.maketrans({"♯": "#", "𝄪": "##", "♭": "b", "𝄫": "bb"})
-        notes = list(n.translate(translations) for n in notes)
-    return notes
